// angular-project/src/app/model/requisito.ts
import CaacLight from "../fede/model/caac-light";

export default class Requisito { //ES UN ASCO, PERO ES LO QUE HAY

  id: number;
  actaasa: boolean;
  altbanca: boolean ;
  cronalimen: boolean ;
  crondiayh: boolean ;
  cronmerc: boolean ;
  cronparad: boolean ;
  cvsliscom: boolean ;
  desaut: boolean ;
  forafip: boolean ;
  fotodni: boolean ;
  lisrrhh: boolean ;
  organi: boolean ;
  perjur: boolean ;
  prodesex: boolean ;
  segmalpra: boolean ;
  segrescivil: boolean ;
  soliform: boolean ;
  statuto: boolean ;
  casa: CaacLight;

}